+++ src/gicv3/distributor.rs
@@ -10,7 +10,7 @@ use crate::{
     },
 };
 use core::{hint::spin_loop, ops::Range};
-use safe_mmio::{UniqueMmioPointer, field, field_shared};
+use safe_mmio::{UniqueMmioPointer, field, field_shared, fields::ReadPureWrite};
 use zerocopy::{transmute_mut, transmute_ref};
 
 /// Selects regular or extended registers based on in the `IntId`.
@@ -318,20 +318,56 @@ impl<'a> GicDistributor<'a> {
         );
 
         // Setup the default (E)SPI priorities.
-        set_regs(
-            field!(self.regs, ipriorityr),
-            SPI_START,
-            spi_count,
-            Gicd::IPRIORITY_BITS,
+
+        // For performance reasons this array of registers is initialized as words, so four
+        // priority values can be set in one iteration.
+        let ipriority_word_value = u32::from_le_bytes([
             HIGHEST_NS_PRIORITY,
-        );
+            HIGHEST_NS_PRIORITY,
+            HIGHEST_NS_PRIORITY,
+            HIGHEST_NS_PRIORITY,
+        ]);
+
+        let mut ipriorityr_bytes = field!(self.regs, ipriorityr);
+
+        // Safety: ipriorityr_bytes is a valid pointer to the ipriority register array. Section
+        // 12.1.3 GIC memory-mapped register access of the GIC specificition describes that
+        // GICD_IPRIORITYR supports both 8-bit and 32-bit accesses.
+        let ipriority_words = unsafe {
+            UniqueMmioPointer::new(
+                ipriorityr_bytes
+                    .ptr_nonnull()
+                    .cast::<[ReadPureWrite<u32>; 1024 / 4]>(),
+            )
+        };
 
         set_regs(
-            field!(self.regs, ipriorityr_e),
+            ipriority_words.into(),
+            SPI_START / 4,
+            spi_count / 4,
+            Gicd::IPRIORITY_BITS * 4,
+            ipriority_word_value,
+        );
+
+        let mut ipriorityr_e_bytes = field!(self.regs, ipriorityr_e);
+
+        // Safety: ipriorityr_bytes is a valid pointer to the ipriority register array. Section
+        // 12.1.3 GIC memory-mapped register access of the GIC specificition describes that
+        // GICD_IPRIORITYR supports both 8-bit and 32-bit accesses.
+        let ipriority_e_words = unsafe {
+            UniqueMmioPointer::new(
+                ipriorityr_e_bytes
+                    .ptr_nonnull()
+                    .cast::<[ReadPureWrite<u32>; 1024 / 4]>(),
+            )
+        };
+
+        set_regs(
+            ipriority_e_words.into(),
             0,
-            espi_count,
-            Gicd::IPRIORITY_BITS,
-            HIGHEST_NS_PRIORITY,
+            espi_count / 4,
+            Gicd::IPRIORITY_BITS * 4,
+            ipriority_word_value,
         );
 
         // Treat all (E)SPIs as level triggered by default.